requirements.txt
PyQt5
PySide2
QT-PyQt-PySide-Custom-Widgets
pipwin
#pipwin install cairocffi